// Task 7/job listing app/app/models/JobPost.ts
export default interface JobPost {
    id: number;
    title: string;
    description: string;
    responsibilities: string[];
    ideal_candidate: {
      age: string;
      gender: string;
      traits: string[];
    };
    when_where: string;
    about: {
      posted_on: string;
      deadline: string;
      location: string;
      start_date: string;
      end_date: string;
      categories: string[];
      required_skills: string[];
    };
    company: string;
    companyLogoUrl: string;
    image: string;
  }
  